requirements.txt
python-docx
aiohttp 